Make CFile::Seek() take an off_t as param, not an unsigned long
git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1458 726aef4b-f618-498e-8847-2d620e286838

+++ FileUtils.cpp
@@ -236,8 +236,8 @@ bool CFile::Chmod(const CString& sFile, mode_t mode) {
 	return (chmod(sFile.c_str(), mode) == 0);
 }
 
-bool CFile::Seek(unsigned long uPos) {
-	if (m_iFD != -1 && (unsigned int) lseek(m_iFD, uPos, SEEK_SET) == uPos) {
+bool CFile::Seek(off_t uPos) {
+	if (m_iFD != -1 && lseek(m_iFD, uPos, SEEK_SET) == uPos) {
 		ClearBuffer();
 		return true;
 	}
